refactor: extract god-object sub-modules per audit core-arch finding #2
Three first-pass splits along the seams flagged in the v3.10.1 coordinated
audit. Each original file is still over the 500-LOC CLAUDE.md limit and
needs more extractions to follow; this commit establishes the pattern and
lands the cleanest seam from each.

1. commands/hooks.ts (5329 -> 4809 LOC):
   Extracted the worker subcommand cluster (5 worker.* commands + parent
   wrapper + formatWorkerStatus helper) to commands/hooks/worker.ts.
   Main hooks.ts imports `workerCommand` and continues to assemble it
   into the hooksCommand.subcommands array.

2. memory/memory-initializer.ts (2991 -> 2744 LOC):
   Extracted the pure-function vector ops (Int8 quantization +
   Flash-Attention-style batch search / softmax / topK) to
   memory/vector-ops.ts. No external state; re-exported from
   memory-initializer.ts for callers that still import from the old path.

3. mcp-tools/hooks-tools.ts (4188 -> 3584 LOC):
   Extracted all 5 worker MCP tools (hooks_worker-{list,dispatch,status,
   detect,cancel}) plus WORKER_CONFIGS / WORKER_TRIGGER_PATTERNS /
   activeWorkers state / detectWorkerTriggers helper to
   mcp-tools/hooks-worker-tools.ts. hooks-tools.ts imports and
   re-exports the 5 tools so the hooksTools[] array and external
   importers continue to resolve.

Net per-original reduction: 520 / 247 / 604 LOC. New cohesive modules:
525 / 201 / 620 LOC.
